changed to f32 inside vectors

+++ src/vec2.rs
@@ -1,11 +1,11 @@
 #[derive(Debug, Clone, Copy, PartialEq)]
 pub struct Vec2 {
-    pub x: i32,
-    pub y: i32,
+    pub x: f32,
+    pub y: f32,
 }
 
 impl Vec2 {
-    pub fn new(x: i32, y: i32) -> Self {
+    pub fn new(x: f32, y: f32) -> Self {
         Self { x, y }
     }
